requirements.txt
pandas
numpy
requests
pyyaml
pandas_ta
yfinance
alpha_vantage
newsapi-python
ta-lib  # optional: can use pandas_ta if ta-lib is painful to build
python-dateutil
tqdm
